refactor(sidebar-layout): reestructuracion de carpetas y layout
