+++ app/(main)/layout.tsx
@@ -1,5 +1,4 @@
 import { Header } from "@/components/ui/header/Header";
-import { Footer } from "@/components/ui/footer/Footer";
 
 export default function MainLayout({
 	children,
@@ -10,7 +9,6 @@ export default function MainLayout({
 		<main className="main-container block">
 			<Header />
 			{children}
-			<Footer />
 		</main>
 	);
 }
